Debugging

+++ mrcnn/utils.py
@@ -888,6 +888,6 @@ def denorm_boxes(boxes, shape):
         [N, (y1, x1, y2, x2)] in pixel coordinates
     """
     h, w = shape
-    scale = np.array([h - 1, w - 1, h - 1, w - 1])
-    shift = np.array([0, 0, 1, 1])
-    return np.around(np.multiply(boxes, scale) + shift).astype(np.int32)
+    scale = np.array([h - 1, w - 1, h - 1, w - 1]).astype(np.float32)
+    shift = np.array([0, 0, 1, 1]).astype(np.float32)
+    return np.around(np.multiply(boxes.astype(np.float32), scale) + shift).astype(np.int32)
